- Reorder code blocks
- Rename cssClassCollectorWriter to htmlElementCollectorWriter, as it just collects html element information
- Expand benchmark to test for minified and unminified content

Fixes #8396, Fixes #8417