@dirkolbrich 2021-04-17
I even managed to make it slightly faster. Check branch [htmlcollector-bench](https://github.com/gohugoio/hugo/compare/master...dirkolbrich:htmlcollector-bench) for an updated benchmark test on master to make it comparable.

```bash
➜  publisher git:(htmlcollector-bench) gobench --package ../publisher | tee ~/old.txt
Benchmark branch "htmlcollector-bench"

 go version go1.16.3 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/publisher
cpu: Intel(R) Core(TM) i7-6920HQ CPU @ 2.90GHz
BenchmarkClassCollectorWriter-8                            16903             69535 ns/op          103649 B/op        400 allocs/op
BenchmarkClassCollectorWriterMinified-8                    17094             75559 ns/op          104328 B/op        401 allocs/op
BenchmarkClassCollectorWriterWithMinifyStream-8            13074             90168 ns/op          106299 B/op        412 allocs/op
BenchmarkClassCollectorWriterWithMinifyString-8            13304             90376 ns/op          109320 B/op        420 allocs/op
PASS
ok      github.com/gohugoio/hugo/publisher      8.366s
➜  publisher git:(htmlcollector-bench) gobench --package ../publisher | tee ~/new.txt
Benchmark branch "writestats-exclude-comments-and-doctype"

 go version go1.16.3 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/publisher
cpu: Intel(R) Core(TM) i7-6920HQ CPU @ 2.90GHz
BenchmarkClassCollectorWriter-8                            18169             65161 ns/op           96928 B/op        335 allocs/op
BenchmarkClassCollectorWriterMinified-8                    18256             65314 ns/op           97594 B/op        337 allocs/op
BenchmarkClassCollectorWriterWithMinifyStream-8            13780             87129 ns/op           99536 B/op        348 allocs/op
BenchmarkClassCollectorWriterWithMinifyString-8            14005             85706 ns/op          102549 B/op        356 allocs/op
PASS
ok      github.com/gohugoio/hugo/publisher      8.109s
➜  publisher git:(writestats-exclude-comments-and-doctype) benchstat ~/old.txt ~/new.txt 
name                                    old time/op    new time/op    delta
ClassCollectorWriter-8                    69.5µs ± 0%    65.2µs ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterMinified-8            75.6µs ± 0%    65.3µs ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterWithMinifyStream-8    90.2µs ± 0%    87.1µs ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterWithMinifyString-8    90.4µs ± 0%    85.7µs ± 0%   ~     (p=1.000 n=1+1)

name                                    old alloc/op   new alloc/op   delta
ClassCollectorWriter-8                     104kB ± 0%      97kB ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterMinified-8             104kB ± 0%      98kB ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterWithMinifyStream-8     106kB ± 0%     100kB ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterWithMinifyString-8     109kB ± 0%     103kB ± 0%   ~     (p=1.000 n=1+1)

name                                    old allocs/op  new allocs/op  delta
ClassCollectorWriter-8                       400 ± 0%       335 ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterMinified-8               401 ± 0%       337 ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterWithMinifyStream-8       412 ± 0%       348 ± 0%   ~     (p=1.000 n=1+1)
ClassCollectorWriterWithMinifyString-8       420 ± 0%       356 ± 0%   ~     (p=1.000 n=1+1)
```